fix bug that type of time and that filter out zero points and that pointcloud  header error

+++ velodyne_pointcloud/include/velodyne_pointcloud/point_types.h
@@ -57,7 +57,7 @@ POINT_CLOUD_REGISTER_POINT_STRUCT(velodyne_pointcloud::PointXYZIRT,
                                   (float, z, z)
                                   (float, intensity, intensity)
                                   (uint16_t, ring, ring)
-                                  (uint16_t, time, time))
+                                  (double, time, time))
 
 #endif // __VELODYNE_POINTCLOUD_POINT_TYPES_H
